backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
python-dateutil>=2.9.0
alembic>=1.13
openpyxl>=3.1.0
python-docx>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
